move MAB model to own file
